+++ src/plantuml_gui/static/sequence.js
@@ -65,6 +65,34 @@ function sequenceEventListeners() {
     });
 
 
+    document.getElementById('renameParticipant').addEventListener('click', async () => {
+        const element = document.getElementById('colb');
+        const svg = element.querySelector('g');
+        try {
+            const plantuml = trimlines(editor.session.getValue());
+            const response = await fetch("getParticipantName", {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    'plantuml': plantuml,
+                    'svg': svg.innerHTML,
+                    'svgelement': lastclickedsvgelement.outerHTML
+                })
+            });
+            const name = await response.text();
+            $('#participant-name-modalForm .modal-title').text('Rename ' + name);
+            $('#participant-name-text').val(name);
+            $('#participant-name-modalForm').modal('show');
+            $('#participant-name-modalForm').on('shown.bs.modal', function() {
+                $('#participant-name-text').trigger('focus');
+            });
+        } catch (error) {
+            displayErrorMessage(`Error with fetch API: ${error.message}`, error);
+        }
+    });
+
     const sequenceList = [{
         id: 'addParticipantLeft',
         endpoint: 'addParticipant',